[#262] feat: 내 정보 입력 뷰 수정 및 정보 업데이트 하도록 로직 수정
