+++ lib/presentation/entrance/model/model.dart
@@ -1,3 +1,3 @@
 export 'log_in_view_model.dart';
 export 'sign_up_auth_data_view_model.dart';
-export 'sign_up_user_data_view_model.dart';
+export 'edit_user_detail_view_model.dart';
