+++ lib/presentation/entrance/provider/provider.dart
@@ -1,5 +1,5 @@
 export 'auth_notifier.dart';
 export 'auth_state.dart';
+export 'edit_user_detail_view_provider.dart';
 export 'log_in_view_model_provider.dart';
 export 'sign_up_auth_data_view_model_provider.dart';
-export 'sign_up_user_data_view_model_provider.dart';
