+++ lib/presentation/entrance/view/sign_up_view.dart
@@ -7,7 +7,6 @@ import 'package:wehavit/dependency/presentation/viewmodel_dependency.dart';
 import 'package:wehavit/domain/entities/entities.dart';
 import 'package:wehavit/presentation/common_components/common_components.dart';
 import 'package:wehavit/presentation/entrance/entrance.dart';
-import 'package:wehavit/presentation/main/main.dart';
 
 class SignUpAuthDataView extends ConsumerStatefulWidget {
   const SignUpAuthDataView({super.key});
@@ -326,7 +325,9 @@ class _SignUpAuthDataViewState extends ConsumerState<SignUpAuthDataView> {
                           context,
                           MaterialPageRoute(
                             builder: (context) {
-                              return const SignUpUserDetailView();
+                              return const EditUserDetailView(
+                                isModifying: false,
+                              );
                             },
                           ),
                         );
@@ -350,343 +351,6 @@ class _SignUpAuthDataViewState extends ConsumerState<SignUpAuthDataView> {
           ),
         ],
       ),
-    );
-  }
-}
-
-class SignUpUserDetailView extends ConsumerStatefulWidget {
-  const SignUpUserDetailView({super.key});
-
-  @override
-  ConsumerState<SignUpUserDetailView> createState() =>
-      _SignUpUserDetailViewState();
-}
-
-class _SignUpUserDetailViewState extends ConsumerState<SignUpUserDetailView> {
-  @override
-  Widget build(BuildContext context) {
-    final viewmodel = ref.watch(signUpUserDataViewModelProvider);
-    final provider = ref.read(signUpUserDataViewModelProvider.notifier);
-
-    return Stack(
-      children: [
-        Scaffold(
-          resizeToAvoidBottomInset: true,
-          backgroundColor: CustomColors.whDarkBlack,
-          appBar: WehavitAppBar(
-            title: '회원가입',
-            leadingTitle: '',
-            leadingIcon: Icons.chevron_left,
-            leadingAction: () async {
-              try {
-                await provider.removeUserData();
-                await provider.logOut();
-              } on Exception catch (e) {
-                // ignore: avoid_print
-                print('DEBUG: ${e.toString()}');
-              }
-
-              // ignore: use_build_context_synchronously
-              Navigator.pop(context);
-            },
-          ),
-          body: SafeArea(
-            minimum: const EdgeInsets.all(16.0),
-            maintainBottomViewPadding: true,
-            child: Column(
-              children: [
-                Expanded(
-                  child: ListView(
-                    shrinkWrap: true,
-                    padding: const EdgeInsets.only(bottom: 32.0),
-                    keyboardDismissBehavior:
-                        ScrollViewKeyboardDismissBehavior.onDrag,
-                    children: [
-                      TextButton(
-                        style: TextButton.styleFrom(
-                          overlayColor: CustomColors.whYellow,
-                        ),
-                        onPressed: () async {
-                          provider.pickProfileImage().whenComplete(
-                                () => setState(() {}),
-                              );
-                        },
-                        child: Stack(
-                          clipBehavior: Clip.none,
-                          children: [
-                            Container(
-                              width: 85,
-                              height: 85,
-                              decoration: const BoxDecoration(
-                                shape: BoxShape.circle,
-                                color: CustomColors.whGrey,
-                              ),
-                              clipBehavior: Clip.hardEdge,
-                              child: viewmodel.profileImageFile != null
-                                  ? Image.file(
-                                      viewmodel.profileImageFile!,
-                                      fit: BoxFit.cover,
-                                    )
-                                  : Container(),
-                            ),
-                            Positioned(
-                              bottom: 0,
-                              right: -5,
-                              child: Container(
-                                width: 25,
-                                height: 25,
-                                decoration: const BoxDecoration(
-                                  shape: BoxShape.circle,
-                                  color: CustomColors.whWhite,
-                                ),
-                                child: const Icon(
-                                  Icons.photo_camera,
-                                  color: CustomColors.whBlack,
-                                  size: 18.0,
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                      const SizedBox(
-                        height: 24,
-                      ),
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          const Text(
-                            '이름',
-                            style: TextStyle(
-                              fontWeight: FontWeight.w600,
-                              color: CustomColors.whWhite,
-                              fontSize: 20,
-                            ),
-                          ),
-                          Container(
-                            height: 8.0,
-                          ),
-                          TextFormField(
-                            onChanged: (value) {
-                              provider.setName(value);
-                              setState(() {});
-                            },
-                            cursorColor: CustomColors.whWhite,
-                            textAlignVertical: TextAlignVertical.center,
-                            style: const TextStyle(
-                              color: CustomColors.whWhite,
-                              fontSize: 16.0,
-                            ),
-                            decoration: InputDecoration(
-                              hintText: '친구들에게 보여지는 이름이예요',
-                              hintStyle: const TextStyle(
-                                fontSize: 16,
-                                color: CustomColors.whPlaceholderGrey,
-                              ),
-                              filled: true,
-                              fillColor: CustomColors.whGrey,
-                              border: OutlineInputBorder(
-                                borderRadius: BorderRadius.circular(10),
-                                borderSide: const BorderSide(
-                                  width: 0,
-                                  style: BorderStyle.none,
-                                ),
-                              ),
-                              isCollapsed: true,
-                              contentPadding: const EdgeInsets.symmetric(
-                                vertical: 12.0,
-                                horizontal: 16.0,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                      const SizedBox(
-                        height: 24,
-                      ),
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          const Text(
-                            '사용자 ID',
-                            style: TextStyle(
-                              fontWeight: FontWeight.w600,
-                              color: CustomColors.whWhite,
-                              fontSize: 20,
-                            ),
-                          ),
-                          Container(
-                            height: 8.0,
-                          ),
-                          TextFormField(
-                            onChanged: (value) {
-                              provider.setHandle(value);
-                              setState(() {});
-                            },
-                            cursorColor: CustomColors.whWhite,
-                            textAlignVertical: TextAlignVertical.center,
-                            inputFormatters: <TextInputFormatter>[
-                              FilteringTextInputFormatter.allow(
-                                RegExp(r'[0-9a-zA-Z!@#$%^&*(),.?":{}|<>_]'),
-                              ),
-                            ],
-                            style: const TextStyle(
-                              color: CustomColors.whWhite,
-                              fontSize: 16.0,
-                            ),
-                            decoration: InputDecoration(
-                              hintText: '친구가 나를 찾을 때 사용하는 ID예요',
-                              hintStyle: const TextStyle(
-                                fontSize: 16,
-                                color: CustomColors.whPlaceholderGrey,
-                              ),
-                              filled: true,
-                              fillColor: CustomColors.whGrey,
-                              border: OutlineInputBorder(
-                                borderRadius: BorderRadius.circular(10),
-                                borderSide: const BorderSide(
-                                  width: 0,
-                                  style: BorderStyle.none,
-                                ),
-                              ),
-                              isCollapsed: true,
-                              contentPadding: const EdgeInsets.symmetric(
-                                vertical: 12.0,
-                                horizontal: 16.0,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                      const SizedBox(
-                        height: 24,
-                      ),
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          const Text(
-                            '한 줄 소개',
-                            style: TextStyle(
-                              fontWeight: FontWeight.w600,
-                              color: CustomColors.whWhite,
-                              fontSize: 20,
-                            ),
-                          ),
-                          Container(
-                            height: 8.0,
-                          ),
-                          TextFormField(
-                            onChanged: (value) {
-                              provider.setAboutMe(value);
-                              setState(() {});
-                            },
-                            cursorColor: CustomColors.whWhite,
-                            textAlignVertical: TextAlignVertical.center,
-                            style: const TextStyle(
-                              color: CustomColors.whWhite,
-                              fontSize: 16.0,
-                            ),
-                            decoration: InputDecoration(
-                              hintText: '나에 대해 소개해주세요',
-                              hintStyle: const TextStyle(
-                                fontSize: 16,
-                                color: CustomColors.whPlaceholderGrey,
-                              ),
-                              filled: true,
-                              fillColor: CustomColors.whGrey,
-                              border: OutlineInputBorder(
-                                borderRadius: BorderRadius.circular(10),
-                                borderSide: const BorderSide(
-                                  width: 0,
-                                  style: BorderStyle.none,
-                                ),
-                              ),
-                              isCollapsed: true,
-                              contentPadding: const EdgeInsets.symmetric(
-                                vertical: 12.0,
-                                horizontal: 16.0,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                      // Expanded(child: Container()),
-                    ],
-                  ),
-                ),
-                WideColoredButton(
-                  onPressed: () async {
-                    setState(() {
-                      viewmodel.isProcessing = true;
-                    });
-
-                    await provider.registerUserData().then(
-                          (result) => result.fold(
-                            (failure) {
-                              String toastMessage = '';
-                              switch (failure.message) {
-                                case 'handle-already-exist':
-                                  toastMessage = '이미 사용중인 ID예요';
-                                  break;
-                                case 'no-image-file':
-                                  toastMessage = '프로필 이미지를 업로드해주세요';
-                                  break;
-                                case 'no-handle':
-                                  toastMessage = '사용자 ID를 업로드해주세요';
-                                  break;
-                                default:
-                                  toastMessage = '잠시 후 다시 시도해주세요';
-                                  break;
-                              }
-
-                              showToastMessage(
-                                context,
-                                text: toastMessage,
-                                icon: const Icon(
-                                  Icons.not_interested,
-                                  color: PointColors.red,
-                                ),
-                              );
-                            },
-                            (success) async {
-                              Navigator.pushReplacement(
-                                context,
-                                MaterialPageRoute(
-                                  fullscreenDialog: true,
-                                  builder: (context) => const MainView(),
-                                ),
-                              );
-                            },
-                          ),
-                        );
-                    setState(() {
-                      viewmodel.isProcessing = false;
-                    });
-                  },
-                  isDiminished: !((viewmodel.name.isNotEmpty &
-                              viewmodel.handle.isNotEmpty) &
-                          (viewmodel.profileImageFile != null)) |
-                      viewmodel.isProcessing,
-                  buttonTitle: viewmodel.isProcessing ? '처리 중' : '완료',
-                  backgroundColor: CustomColors.whYellow,
-                  foregroundColor: CustomColors.whBlack,
-                ),
-              ],
-            ),
-          ),
-        ),
-        Visibility(
-          visible: viewmodel.isProcessing,
-          child: Container(
-            constraints: const BoxConstraints.expand(),
-            alignment: Alignment.center,
-            color: CustomColors.whDarkBlack.withAlpha(130),
-            child: const CircularProgressIndicator(
-              color: CustomColors.whYellow,
-            ),
-          ),
-        ),
-      ],
     );
   }
 }
