+++ lib/presentation/entrance/view/view.dart
@@ -1,3 +1,4 @@
+export 'edit_user_detail_view.dart';
 export 'entrance_view.dart';
 export 'log_in_view.dart';
 export 'sign_up_view.dart';
